style(donation-machines): rename machine folder to plural
as this folder will be intended to expose all machines needed for a given donation flow
